# Cache "global" trait resolutions