@arielb1 2015-06-15
Translation creates tons of these useless nested obligations - I think the primary benefits may be there. Could you post some performance numbers?