@arielb1 2015-06-16
Anyway, rustc perf numbers:

```
Before:
548.54user 4.54system 7:16.79elapsed 126%CPU (0avgtext+0avgdata 1174632maxresident)k

After:
537.65user 4.22system 7:04.81elapsed 127%CPU (0avgtext+0avgdata 1175460maxresident)k
```

This is a 2%-3% improvement on rustc. It is indeed mostly a trans improvement (improves translation from 89.6s to 79.3s, but type-checking from 62.6s to 59s) - probably because of workload differences between rustc and trans. Still nice.