@nrc 2015-06-16
r=me with the rebasing errors fixed