@nikomatsakis 2015-06-16
@arielb1 time numbers for compiling script (in dev mode) on my machine:

**After:**

- real    2m2.024s
- user    1m58.926s
- sys     0m3.098s

**Before:**

- real    2m35.145s
- user    2m32.113s
- sys     0m2.980s

Regarding error messages, I discussed it with @pcwalton, and we felt like it was worth landing this sooner and dealing with error messages later -- though I left a note for what I think is the best fix. Doesn't seem *too* hard, I could do some local experiments.