@arielb1 2015-06-16
This does indeed seem to produce 13-14% improvements in both typeck and trans time when assertions are enabled (I guess these `debug!`-s add up).